refactor: Update functions to use python-binance package
